pyproject.toml
[project]
name = "arrlp"
version = "0.1.16"
description = "A library that provides tools for arrays on CPU & GPU [any shape]."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [ "corelp", "joblib", "numba", "numpy", "scipy",]

[build-system]
requires = [ "uv_build",]
build-backend = "uv_build"

[dependency-groups]
tests = [ "pytest",]
docs = [ "sphinx", "sphinx-design", "sphinx-rtd-theme", "toml",]
anaconda = [ "ipykernel", "spyder-kernels",]

[project.optional-dependencies]
cuda11 = [ "cupy-cuda11x",]
cuda12 = [ "cupy-cuda12x",]
cuda13 = [ "cupy-cuda13x",]

[tool.uv]
required-environments = [ "sys_platform == 'linux' and platform_machine == 'x86_64'", "sys_platform == 'win32' and (platform_machine == 'AMD64' or platform_machine == 'x86_64')",]

[tool.uv.sources.corelp]
workspace = true
